move: ZoomOverlay 컴포넌트 분리
RestaurantOverlay -> ClickOverlay 컴포넌트 이름 변경
